+++ packages/core/src/lib/shared.ts
@@ -1,3 +1,3 @@
-import type { Context } from "./context";
+import type { Context } from "./context.js";
 
 export type ConstOrDynamic<T> = T | ((context: Context) => T | Promise<T>);