+++ packages/core/src/types.ts
@@ -1,6 +1,6 @@
-export type { CLI, SelectOption } from "./lib/cli";
-export type { Context } from "./lib/context";
-export type { Plugin } from "./lib/plugin";
-export type { ConstOrDynamic } from "./lib/shared";
-export type { Stage } from "./lib/stage";
-export type { System } from "./lib/system";
+export type { CLI, SelectOption } from "./lib/cli.js";
+export type { Context } from "./lib/context.js";
+export type { Plugin } from "./lib/plugin.js";
+export type { ConstOrDynamic } from "./lib/shared.js";
+export type { Stage } from "./lib/stage.js";
+export type { System } from "./lib/system.js";