+++ packages/plugin-license/src/index.test.ts
@@ -1,5 +1,6 @@
 import { DefaultStages } from "@alcalzone/release-script-core";
 import { createMockContext, TestFS } from "@alcalzone/release-script-testing";
+import { afterEach, beforeEach, describe, expect, it } from "vitest";
 import LicensePlugin from ".";
 
 describe("License plugin", () => {