+++ packages/release-script/bin/release.js
@@ -1,2 +1,2 @@
 #!/usr/bin/env node
-require("../build/index.js");
+import "../build/index.js";